rename data fill to clean_data.py
